+++ src/libs/SlowTicker.h
@@ -30,7 +30,7 @@ class SlowTicker : public Module{
         // TODO replace this with std::function()
         template<typename T> Hook* attach( uint32_t frequency, T *optr, uint32_t ( T::*fptr )( uint32_t ) ){
             Hook* hook = new Hook();
-            hook->interval = 1/frequency;
+            hook->interval = ((float)1)/frequency;
             hook->attach(optr, fptr);
             hook->countdown = hook->interval;
